requirements.txt
# Core Dependencies
streamlit>=1.28.0
langgraph>=0.2.45
pyyaml>=6.0
python-dotenv>=1.0.0

# LLM Providers
groq>=1.0.0
ollama>=0.6.0

# Visualization
plotly>=5.14.0
pyvis>=0.3.2

# Database (PostgreSQL for production)
psycopg2-binary>=2.9.0

# Utilities
requests>=2.31.0
